all INIT permission are added and Autorisations

+++ backend/src/main/java/User/Recht/Tool/resource/PermissionResource.java
@@ -2,14 +2,20 @@ package User.Recht.Tool.resource;
 
 import User.Recht.Tool.dtos.permissionDtos.PermissionDto;
 import User.Recht.Tool.entity.Permission;
+import User.Recht.Tool.entity.User;
+import User.Recht.Tool.exception.Permission.CannotDeleteInitPermissions;
 import User.Recht.Tool.exception.Permission.PermissionNotFound;
 import User.Recht.Tool.exception.Permission.PermissionToRoleNotFound;
+import User.Recht.Tool.exception.Permission.UserNotAuthorized;
 import User.Recht.Tool.exception.role.RoleNotFoundException;
 import User.Recht.Tool.exception.superadmin.CannotModifySuperAdminException;
+import User.Recht.Tool.exception.user.UserNotFoundException;
+import User.Recht.Tool.service.AutorisationService;
 import User.Recht.Tool.service.PermissionService;
+import User.Recht.Tool.service.UserService;
+import io.vertx.ext.web.RoutingContext;
 import org.eclipse.microprofile.openapi.annotations.parameters.RequestBody;
 
-import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
 import javax.inject.Inject;
 import javax.ws.rs.*;
@@ -26,13 +32,23 @@ public class PermissionResource {
 
     @Inject
     PermissionService permissionService;
+    @Inject
+    UserService userService;
+    @Inject
+    AutorisationService autorisationService;
 
     @POST
-    @RolesAllowed({ "USER" })
-    public Response addPermissions(@RequestBody PermissionDto permissionDto, @Context SecurityContext securityContext) {
+    @RolesAllowed({"USER"})
+    public Response addPermissions(@RequestBody PermissionDto permissionDto,
+                                   @Context RoutingContext routingContext, @Context SecurityContext securityContext) {
 
         try {
-            List <Permission> permissions = permissionService.createPermission(permissionDto);
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_POST", token);
+
+            List<Permission> permissions = permissionService.createPermission(permissionDto);
             return Response.ok(permissions).header("status", "NEW PERMISSIONS ARE ADDED")
                     .build();
 
@@ -53,36 +69,87 @@ public class PermissionResource {
                     .header("status", "xxx")
                     .build();
 
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
         }
     }
 
     @GET
-    @RolesAllowed({ "USER" })
-    public Response getAllPermissions(@Context SecurityContext securityContext) {
-
-        List<Permission> permissions = permissionService.getAllPermissions();
-
-        return Response.ok(permissions).header("STATUS", "LIST OF PERMISSIONS")
-                .build();
-    }
-    @GET
-    @RolesAllowed({ "USER" })
-    @Path("name/{name}")
-    public Response getPermissionsByName(@PathParam("name") String name,@Context SecurityContext securityContext) {
-
-        List<Permission> permissions = permissionService.getPermissionsByName(name);
-
-        return Response.ok(permissions).header("STATUS", "LIST OF PERMISSIONS")
-                .build();
-    }
-
-    @GET
-    @RolesAllowed({ "USER" })
-    @Path("key/{key}")
-    public Response getPermissionByKey(@PathParam("key") String key,@Context SecurityContext securityContext) {
-
+    @RolesAllowed({"USER"})
+    public Response getAllPermissions(@Context RoutingContext routingContext, @Context SecurityContext securityContext) {
 
         try {
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_GET", token);
+            List<Permission> permissions = permissionService.getAllPermissions();
+
+            return Response.ok(permissions).header("STATUS", "LIST OF PERMISSIONS")
+                    .build();
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
+        }
+
+
+    }
+
+    @GET
+    @RolesAllowed({"USER"})
+    @Path("name/{name}")
+    public Response getPermissionsByName(@PathParam("name") String name,
+                                         @Context RoutingContext routingContext, @Context SecurityContext securityContext) {
+
+        try {
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_GET", token);
+
+            List<Permission> permissions = permissionService.getPermissionsByName(name);
+            return Response.ok(permissions).header("STATUS", "LIST OF PERMISSIONS")
+                    .build();
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
+        }
+    }
+
+    @GET
+    @RolesAllowed({"USER"})
+    @Path("key/{key}")
+    public Response getPermissionByKey(@PathParam("key") String key,
+                                       @Context RoutingContext routingContext, @Context SecurityContext securityContext) {
+
+        try {
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_GET", token);
+
             Permission permission = permissionService.getPermissionByKey(key);
 
             return Response.ok(permission).header("PERMISSION_KEY", key)
@@ -91,15 +158,30 @@ public class PermissionResource {
 
             return Response.status(406, "PERMISSION DONT EXIST")
                     .header("status", "PERMISSION DONT EXIST").build();
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
         }
     }
 
     @DELETE
     @RolesAllowed({ "USER" })
     @Path("key/{key}")
-    public Response deletePermission(@PathParam("key") String key, @Context SecurityContext securityContext) {
+    public Response deletePermission(@PathParam("key") String key, @Context RoutingContext routingContext, @Context SecurityContext securityContext) {
 
         try {
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_DELETE", token);
+
             Permission permissionToDelete = permissionService.deletePermissionByKey(key);
 
             return Response.ok(permissionToDelete).header("STATUS", "THE PERMISSION_KEY " + key + " IS DELETED")
@@ -119,13 +201,34 @@ public class PermissionResource {
         } catch (IllegalArgumentException e) {
             return Response.status(406, "TYPE SHOULD BE ALL OR ONE")
                     .header("status", "TYPE SHOULD BE ALL OR ONE").build();
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
+        }catch (CannotDeleteInitPermissions e) {
+
+            return Response.status(406, "CANNOT DELETE AN INITIALE PERMISSION")
+                    .header("STATUS", "CANNOT DELETE AN INITIALE PERMISSION").build();
+
         }
     }
     @DELETE
     @RolesAllowed({ "USER" })
     @Path("name/{name}")
-    public Response deletePermissionsByName(@PathParam("name") String name, @Context SecurityContext securityContext) {
+    public Response deletePermissionsByName(@PathParam("name") String name, @Context RoutingContext routingContext, @Context SecurityContext securityContext) {
+
         try {
+            User connectedUser = userService.getUserByEmail(securityContext.getUserPrincipal().getName());
+            String token = routingContext.request().getHeader("Authorization").substring(7);
+            // CHECK PERMISSIONS
+            autorisationService.checkExistedUserPermission("PERMISSION_MANAGER_DELETE", token);
+
             List<Permission> permissionToDelete = permissionService.deletePermissionsByName(name);
 
             return Response.ok(permissionToDelete).header("STATUS", "THE PERMISSIONS WITH THE NAME  " + name + " are DELETED")
@@ -145,6 +248,21 @@ public class PermissionResource {
         } catch (IllegalArgumentException e) {
             return Response.status(406, "TYPE SHOULD BE ALL OR ONE")
                     .header("status", "TYPE SHOULD BE ALL OR ONE").build();
+        } catch (UserNotFoundException e) {
+
+            return Response.status(406, "USER DOSENT EXIST")
+                    .header("status", "USER DOSENT EXIST").build();
+
+        } catch (UserNotAuthorized e) {
+
+            return Response.status(406, "USER IS NOT AUTHOROZIED FOR THE PERMISSION")
+                    .header("STATUS", "USER IS NOT AUTHOROZIED FOR THE PERMISSION").build();
+
+        }catch (CannotDeleteInitPermissions e) {
+
+            return Response.status(406, "CANNOT DELETE AN INITIALE PERMISSION")
+                    .header("STATUS", "CANNOT DELETE AN INITIALE PERMISSION").build();
+
         }
     }
 }
